app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
boto3==1.29.3
pydantic==2.5.2
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.23 